Migrate tests 1500-2000
